requirements.txt
pytorch
spacy
nltk
numpy
tqdm
scikit-learn
scipy
sklearn
pytest
pytest-xdist
pytest-forked